Create hex-converter-08-12-21
Finished December 8th, 2021. Thought making this would help solidify my understanding of converting to hex. At the time I was also trying to figure out how to iterate through dictionaries and comment code, but clearly hadn't learned of any style guides.  In any case, this is a snapshot of the what I knew at the given time. While it works, its purpose is more historical for the purpose of tracking progress.